Update pca.ipynb
I hope the change I am suggesting is correct, but if I was able to follow the paragraph correctly, it should be. I got confused a lot by the last sentence in the paragraph : This gives you a score for each ...  ...also scaled by the singular value S1. Because suddenly v(1) appeared in that sentence and also S1 was mentioned, but in a way that suggests it was mentioned before in that paragraph. I believe my suggestion would make the pattern of the sentences more clear. 

Best regards,
Lennart Pielström